Add draw controls to core Map, fix plot  and collect sample widgets
